pyproject.toml
[tool.poetry]
name = "my-first-langgraph"
version = "0.1.0"
description = ""
authors = ["abe.akira"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.2.15"
httpx = "^0.27.2"
langchain-openai = "^0.1.23"
ipython = "^8.27.0"
graphviz = "^0.20.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
